refactor: 도메인의 패키지 분리
