+++ src/main/java/kitchenpos/application/dto/request/OrderRequest.java
@@ -3,8 +3,8 @@ package kitchenpos.application.dto.request;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
 
 public class OrderRequest {
