+++ src/main/java/kitchenpos/dao/JdbcTemplateTableGroupDao.java
@@ -16,7 +16,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.table.TableGroup;
 
 @Repository
 public class JdbcTemplateTableGroupDao implements TableGroupDao {
