+++ src/main/java/kitchenpos/domain/menu/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import java.util.List;
 import java.util.Optional;
