+++ src/main/java/kitchenpos/repository/OrderRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderLineItemDao;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
 
 @Repository
 public class OrderRepository {
