+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -15,11 +15,11 @@ import org.springframework.transaction.annotation.Transactional;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.TableGroupDao;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.TableGroup;
 import kitchenpos.application.dto.request.ChangeNumOfTableGuestsRequest;
 import kitchenpos.application.dto.request.ChangeOrderTableEmptyRequest;
 import kitchenpos.application.dto.request.OrderTableRequest;
