+++ src/test/java/kitchenpos/domain/MenuTest.java
@@ -9,6 +9,11 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuGroup;
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.Product;
+
 class MenuTest {
 
     @DisplayName("모든 메뉴 상품을 추가한다")
