+++ src/test/java/kitchenpos/fixture/MenuFixture.java
@@ -7,10 +7,10 @@ import java.util.stream.Collectors;
 
 import kitchenpos.application.dto.request.MenuProductRequest;
 import kitchenpos.application.dto.request.MenuRequest;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuGroup;
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.Product;
 
 @SuppressWarnings("NonAsciiCharacters")
 public enum MenuFixture {
