+++ src/test/java/kitchenpos/fixture/ProductFixture.java
@@ -2,7 +2,7 @@ package kitchenpos.fixture;
 
 import java.math.BigDecimal;
 
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Product;
 import kitchenpos.application.dto.request.ProductRequest;
 
 @SuppressWarnings("NonAsciiCharacters")
